rename rag state
