+++ lexio/lib/state/rag-state.ts
@@ -1,651 +1,789 @@
-import { atom, WritableAtom } from 'jotai'
-import { acceptsSources, GenerateInput, GenerateResponse, GenerateSimple, GenerateStreamChunk, GenerateWithSources, GetDataSourceResponse, Message, RAGConfig, RAGWorkflowActionOnAddMessage, RetrievalResult, RetrieveAndGenerateResponse, RetrieveResponse, SourceContent, SourceReference, WorkflowMode } from '../types';
-import { toast } from 'react-toastify';
-import { validateRetrievalResults } from './data_validation';
+import { atom } from 'jotai'
 
-export const workflowModeAtom = atom<WorkflowMode>('init');
+// ---- central state management types -----
+export type Component =
+    | 'RAGProvider'
+    | 'QueryField'
+    | 'ChatWindow'
+    | 'SourcesDisplay'
+    | 'AdvancedQueryField'
+    | `QueryField-${string}`
+    | `ChatWindow-${string}`
+    | `SourcesDisplay-${string}`
+    | `AdvancedQueryField-${string}`
 
-export const ragConfigAtom = atom<RAGConfig>({
-  timeouts: {
-    stream: 10000,  // 10 seconds default
-    request: 30000  // 30 seconds default
-  }
+// Flow: action in component -> triggers wrapped user function with state props -> triggers dispatch -> manipulates state
+export type UserAction =
+    | { type: 'ADD_USER_MESSAGE', message: string, source: Component } // User message
+    | { type: 'SET_ACTIVE_MESSAGE', messageId: string, source: Component } // Set active message -> can be used to rollback to a previous message
+    | { type: 'CLEAR_MESSAGES', source: Component } // Clear all messages in the chat
+    | { type: 'SEARCH_SOURCES', query: string, source: Component } // Search sources, retrieve function - triggered by SourceDisplay
+    | { type: 'CLEAR_SOURCES', source: Component } // Clear all sources
+    | { type: 'SET_ACTIVE_SOURCES', sourceIds: string[], source: Component } // Sets active sources
+    | { type: 'SET_SELECTED_SOURCE', sourceId: string, source: Component } // Sets selected source -> viewers
+    | { type: 'SET_FILTER_SOURCES', filter: any, source: Component } // todo: define filter object
+    | { type: 'RESET_FILTER_SOURCES', source: Component }
+
+interface AddUserMessageActionModifier {
+    setUserMessage?: string
+}
+
+interface SetActiveMessageActionModifier {
+    messageId: string
+}
+
+interface ClearMessagesActionModifier {
+
+}
+
+interface SearchSourcesActionModifier {
+
+}
+
+interface ClearSourcesActionModifier {
+
+}
+
+interface SetActiveSourcesActionModifier {
+    activeSourceIds: string[]
+}
+
+interface SetSelectedSourceActionModifier {
+    selectedSourceId: string | null
+}
+
+interface SetFilterSourcesActionModifier {
+
+}
+
+interface ResetFilterSourcesActionModifier {
+
+}
+
+const allowedModifiers: Record<UserAction['type'], object> = {
+    ADD_USER_MESSAGE: { setUserMessage: '' },
+    SET_ACTIVE_MESSAGE: { messageId: '' },
+    CLEAR_MESSAGES: {},
+    SEARCH_SOURCES: {},
+    CLEAR_SOURCES: {},
+    SET_ACTIVE_SOURCES: { activeSourceIds: [] },
+    SET_SELECTED_SOURCE: { selectedSourceId: '' },
+    SET_FILTER_SOURCES: {},
+    RESET_FILTER_SOURCES: {}
+};
+
+const allowedPayloadKeys: Record<UserAction['type'], string[]> = {
+    ADD_USER_MESSAGE: ['response', 'message', 'messages', 'sources', 'actionOptions'],
+    SET_ACTIVE_MESSAGE: ['messageId', 'actionOptions'],
+    CLEAR_MESSAGES: ['actionOptions'],
+    SEARCH_SOURCES: ['sources', 'actionOptions'],
+    CLEAR_SOURCES: ['actionOptions'],
+    SET_ACTIVE_SOURCES: ['sourceIds', 'actionOptions'],
+    SET_SELECTED_SOURCE: ['sourceId', 'actionOptions'],
+    SET_FILTER_SOURCES: ['filter', 'actionOptions'],
+    RESET_FILTER_SOURCES: ['actionOptions']
+};
+
+/**
+ * Represents a highlight annotation in a PDF document.
+ *
+ * @interface PDFHighlight
+ * @property {number} page - The page number where the highlight appears. Page numbers are 1-based.
+ * @property {object} rect - The rectangle coordinates of the highlight
+ * @property {number} rect.top - Top position of the highlight (relative to the page)
+ * @property {number} rect.left - Left position of the highlight (relative to the page)
+ * @property {number} rect.width - Width of the highlight (relative to the page width)
+ * @property {number} rect.height - Height of the highlight (relative to the page height)
+ */
+export interface PDFHighlight {
+    /**
+     * The page number where the highlight appears. Page numbers are 1-based.
+     */
+    page: number;
+    /**
+     * The rectangle coordinates of the highlight
+     */
+    rect: {
+        /**
+         * Top position of the highlight (relative to the page)
+         */
+        top: number;
+        /**
+         * Left position of the highlight (relative to the page)
+         */
+        left: number;
+        /**
+         * Width of the highlight (relative to the page width)
+         */
+        width: number;
+        /**
+         * Height of the highlight (relative to the page height)
+         */
+        height: number;
+    };
+}
+
+export type UUID = `${string}-${string}-${string}-${string}-${string}`;
+
+export interface Message {
+    readonly id: UUID;  // todo: make sure this is unique -> validate changes
+    role: "user" | "assistant";
+    content: string;
+}
+
+export interface Source {
+    readonly id: UUID;  // todo: make sure this is unique -> validate changes
+    title: string;
+    type: "text" | "pdf" | "markdown" | "html";
+    relevance?: number;
+    data?: string | Uint8Array | null;
+    /**
+     * key convention to hide from display _key
+     *
+     * If type='pdf', you can set the 'page' and '_page' properties to specify the page number to display in the ContentDisplay component.
+     */
+    metadata?: Record<string, any> & {
+        /**
+       * The page number to display for PDF documents. Page numbers are 1-based.
+       * @TJS-type integer
+       * @minimum 1
+       */
+        page?: number;
+        /**
+       * Hidden from display. The page number to display for PDF documents. Page numbers are 1-based.
+       * @TJS-type integer
+       * @minimum 1
+       */
+        _page?: number;
+    };
+    /**
+     * Highlight annotations in the PDF document. Only applicable for PDF sources.
+     */
+    highlights?: PDFHighlight[];
+}
+
+// ---- central data atoms -----
+export interface ProviderConfig {
+    timeouts?: {
+        stream?: number;  // Timeout between stream chunks in ms
+        request?: number; // Overall request timeout in ms
+    },
+}
+
+export const configAtom = atom<ProviderConfig>({
+    timeouts: {
+        stream: 2000,  // 2 seconds default
+        request: 5000  // 5 seconds default
+    }
 });
 
-// Current definition (might be wrong)
-export const onAddMessageAtom = atom<((message: Message, previousMessages: Message[]) => RAGWorkflowActionOnAddMessage) | null>(null);
-
+// message, active message, messages, and stream
 export const completedMessagesAtom = atom<Message[]>([]);
 export const currentStreamAtom = atom<Message | null>(null);
+export const activeMessageIdAtom = atom<string | null>(null);
+
+export const activeMessageAtom = atom(
+    (get) => {
+        const activeMessageId = get(activeMessageIdAtom);
+        const completedMessages = get(completedMessagesAtom);
+        return activeMessageId ? completedMessages.find(message => message.id === activeMessageId) ?? null : null;
+    }
+);
+
+
+// sources, active sources, selected source
+export const retrievedSourcesAtom = atom<Source[]>([]);
+export const activeSourcesIdsAtom = atom<string[]>([]);
+export const selectedSourceIdAtom = atom<string | null>(null);
+
+export const activeSourcesAtom = atom(
+    (get) => {
+        const retrievedSources = get(retrievedSourcesAtom);
+        const activeIds = get(activeSourcesIdsAtom);
+        return retrievedSources.filter(source => activeIds.includes(source.id));
+    }
+);
+
+export const selectedSourceAtom = atom(
+    (get) => {
+        const retrievedSources = get(retrievedSourcesAtom);
+        const selectedId = get(selectedSourceIdAtom);
+        return selectedId ? retrievedSources.find(source => source.id === selectedId) ?? null : null;
+    }
+);
+
+// loading state of system -> can be used to block other actions
 export const loadingAtom = atom(false);
+
+// error state of system -> can be used to block other actions
 export const errorAtom = atom<string | null>(null);
 
-
-// Active source index atom
-export const activeSourceIndexAtom = atom<number | null>(null);
-
-// Update setActiveSourceIndexAtom to use getDataSourceAtom directly
-export const setActiveSourceIndexAtom = atom(
-  null,
-  (_get, set, index: number | null) => {
-    set(activeSourceIndexAtom, index);
-    if (index !== null) {
-      const sources = _get(retrievedSourcesAtom);
-      const ragAtoms = _get(ragAtomsAtom);
-      if (ragAtoms?.getDataSourceAtom && sources[index]) {
-        set(ragAtoms.getDataSourceAtom, sources[index]);
-      }
+// centralized error setter
+export const setErrorAtom = atom(
+    null, // read value not needed
+    (_get, set, message: string) => {
+        set(errorAtom, `${new Date().toISOString()}: ${message}`);
     }
-  }
 );
 
-export const currentSourceContentAtom = atom<SourceContent | null>(null);
 
+// ---- ActionHandler Function types -----
+// todo: replace with allowedModifiers
+type UserActionModifier =
+    AddUserMessageActionModifier
+    | SetActiveMessageActionModifier
+    | ClearMessagesActionModifier
+    | SearchSourcesActionModifier
+    | ClearSourcesActionModifier
+    | SetActiveSourcesActionModifier
+    | SetSelectedSourceActionModifier
+    | SetFilterSourcesActionModifier
+    | ResetFilterSourcesActionModifier
 
-// Add message atom
-export const addMessageAtom = atom(
-  null,
-  async (_get, set, message: Message) => {
-    // Validate processing state
-    if (_get(loadingAtom)) {
-      set(errorAtom, 'RAG operation already in progress');
-      return;
+export type ActionHandlerResponse = {
+    response?: Promise<string> | AsyncIterable<{ content: string, done?: boolean }>
+    messages?: Promise<Message[]> | any,
+    sources?: Promise<Source[]>,
+    actionOptions?: {
+        current?: UserActionModifier,
+        followUp?: UserAction,
     }
+}
 
-    // Get current state
-    const currentMode = _get(workflowModeAtom);
-    const ragAtoms = _get(ragAtomsAtom);
-    const onAddMessage = _get(onAddMessageAtom);
-    const previousMessages = _get(completedMessagesAtom);
-    const currentSources = _get(retrievedSourcesAtom);
+export type ActionHandler = {
+    component: Component;
+    handler: (actionHandlerFunction: UserAction, messages: Message[], sources: Source[], activeSources: Source[], selectedSource: Source | null) => ActionHandlerResponse
+}
 
+// ---- handler registry -----
+export const registeredActionHandlersAtom = atom<ActionHandler[]>([]);
 
-    if (!ragAtoms) {
-      throw new Error('RAG atoms not initialized');
+// register a new action handler: handlers should be uniquely identified by component name + key
+export const registerActionHandler = atom(null, (get, set, handler: ActionHandler) => {
+    const existingHandlers = get(registeredActionHandlersAtom);
+    const existingHandler = existingHandlers.find(h => h.component === handler.component);
+    if (existingHandler) {
+        throw new Error(`Handler for component ${handler.component} already registered`);
     }
+    set(registeredActionHandlersAtom, [...get(registeredActionHandlersAtom), handler]);
+})
 
-    // Check available atoms for workflow management
-    const hasGenerateAtom = ragAtoms.generateAtom !== null;
-    const hasRetrieveAndGenerateAtom = ragAtoms.retrieveAndGenerateAtom !== null;
-
-    // Check if we have any generation capability
-    if (!hasRetrieveAndGenerateAtom && !hasGenerateAtom) {
-      toast.error('No generation capability available');
-      throw new Error('No generation capability available');
+// unregister an action handler
+export const unregisterActionHandler = atom(null, (get, set, handler: ActionHandler) => {
+    const existingHandlers = get(registeredActionHandlersAtom);
+    const updatedHandlers = existingHandlers.filter(h => h.component !== handler.component);
+    if (updatedHandlers.length === existingHandlers.length) {
+        throw new Error(`Handler for component ${handler.component} not found`);
     }
+    set(registeredActionHandlersAtom, updatedHandlers);
+})
 
-    // Prepare rollback state
-    const rollbackState = {
-      messages: previousMessages,
-      workflowMode: currentMode,
-      sources: currentSources
-    };
 
-    try {
-      // Handle custom workflow actions if provided
-      if (onAddMessage) {
-        const action = onAddMessage(message, previousMessages);
-        const updatedMessages = action.type === 'reretrieve' && !action.preserveHistory ?
-          [message] :
-          [...previousMessages, message];
+// ---- some dummy for crazy agent message manipulation -----
+// todo: implement for agents later on
+// @ts-ignore
+class CrazyAgentMessageManipulator {
+    private readonly messages: Message[] = [];
 
-        set(completedMessagesAtom, updatedMessages);
-
-        if (action.type === 'reretrieve') {
-          if (!hasRetrieveAndGenerateAtom) {
-            toast.error('Retrieval and generation not available');
-            throw new Error('Retrieval and generation not available');
-          }
-          set(workflowModeAtom, 'reretrieve');
-          set(ragAtoms.retrieveAndGenerateAtom!, updatedMessages, undefined);
-          return;
-        }
-
-        if (action.type === 'follow-up') {
-          if (hasGenerateAtom && currentMode === 'follow-up') {
-            await set(ragAtoms.generateAtom!, updatedMessages);
-            return;
-          }
-          // Fall back to retrieveAndGenerate if not in follow-up mode or generate not available
-          if (hasRetrieveAndGenerateAtom) {
-            set(ragAtoms.retrieveAndGenerateAtom!, updatedMessages, undefined);
-            if (hasGenerateAtom) {
-              set(workflowModeAtom, 'follow-up');
-            }
-            return;
-          }
-          toast.error('No generation capability available');
-          throw new Error('No generation capability available');
-        }
-
-        toast.error('Unexpected workflow action');
-        throw new Error('Unexpected workflow action');
-      }
-
-      // Handle default workflow
-      const updatedMessages = [...previousMessages, message];
-      set(completedMessagesAtom, updatedMessages);
-
-      // If retrieveAndGenerate is available, always use it in init/reretrieve mode
-      if (hasRetrieveAndGenerateAtom && (currentMode === 'init' || currentMode === 'reretrieve')) {
-        set(ragAtoms.retrieveAndGenerateAtom!, updatedMessages, undefined);
-        // Switch to follow-up mode if generate is available, otherwise stay in reretrieve
-        if (hasGenerateAtom) {
-          set(workflowModeAtom, 'follow-up');
-        } else {
-          set(workflowModeAtom, 'reretrieve');
-        }
-        return;
-      }
-
-      // Use generate for follow-up if available
-      if (hasGenerateAtom && currentMode === 'follow-up') {
-        set(ragAtoms.generateAtom!, updatedMessages);
-        return;
-      }
-
-      // If we only have retrieveAndGenerate, use it for everything and stay in reretrieve mode
-      if (hasRetrieveAndGenerateAtom) {
-        set(ragAtoms.retrieveAndGenerateAtom!, updatedMessages, undefined);
-        set(workflowModeAtom, 'reretrieve');
-        return;
-      }
-
-    } catch (err) {
-      console.error("Error in addMessageAtom", err);
-      // On failure, rollback state
-      set(workflowModeAtom, rollbackState.workflowMode);
-      set(completedMessagesAtom, rollbackState.messages);
-      set(retrievedSourcesAtom, rollbackState.sources);
-      throw err;
+    public addMessage(message: Message): void {
+        this.messages.push(message);
     }
-  }
-);
+}
 
-/**
- * Creates an atom that handles the generation process.
- * This atom manages response generation with proper error handling and state management.
- * Supports both streaming and non-streaming responses, with optional source context.
- */
-export const createGenerateAtom = (generateFn: GenerateSimple | GenerateWithSources) => {
-  return atom<null, [GenerateInput], GenerateResponse>(
+// ---- Action Atoms -----
+// handling of user messages
+export const addUserMessageAtom = atom(
     null,
-    (_get, set, messages: GenerateInput): GenerateResponse => {
-      // Initialize state and control variables
-      set(loadingAtom, true);
-      set(errorAtom, null);
-      
-      let aborted = false;
-      const abortController = new AbortController();
-      const config = _get(ragConfigAtom);
-      let accumulatedContent = '';
-
-      // Get response from the provided function, including sources if supported
-      const response = acceptsSources(generateFn)
-        ? (generateFn as GenerateWithSources)(
-            messages, 
-            _get(currentSourcesAtom).length > 0 ? _get(currentSourcesAtom) : _get(retrievedSourcesAtom)
-          )
-        : (generateFn as GenerateSimple)(messages);
-
-      /**
-       * Response Processing
-       * Handles different types of responses (string, Promise, or Stream)
-       * Updates UI state as content becomes available.
-       */
-      const processingPromise = (async () => {
-        try {
-          if (typeof response === 'string' || response instanceof Promise) {
-            // Handle direct string or Promise<string> response
-            const content = await addTimeout(
-              response as Promise<string>,
-              config.timeouts?.request,
-              'Response timeout exceeded',
-              abortController.signal
-            );
-            
-            // Update message history with complete response
-            set(currentStreamAtom, { role: 'assistant', content });
-            set(completedMessagesAtom, prev => [...prev, { role: 'assistant', content }]);
-            set(currentStreamAtom, null);
-          } else if (Symbol.asyncIterator in response) {
-            // Handle streaming response
-            const processStream = async () => {
-              const streamTimeout = new StreamTimeout(config.timeouts?.stream);
-              
-              for await (const chunk of response as AsyncIterable<GenerateStreamChunk>) {
-                if (aborted || abortController.signal.aborted) {
-                  break;
-                }
-
-                // Check for chunk timeout and update content
-                streamTimeout.check();
-                accumulatedContent += chunk.content;
-                
-                // Update UI with current stream
-                set(currentStreamAtom, {
-                  role: 'assistant',
-                  content: accumulatedContent
-                });
-
-                if (chunk.done) {
-                  // On completion, update message history
-                  set(completedMessagesAtom, prev => [...prev, {
-                    role: 'assistant',
-                    content: accumulatedContent
-                  }]);
-                  set(currentStreamAtom, null);
-                }
-              }
-            };
-
-            // Add overall timeout for stream processing
-            await addTimeout(
-              processStream(),
-              config.timeouts?.request,
-              'Stream processing timeout exceeded',
-              abortController.signal
-            );
-          } else {
-            throw new Error('Invalid GenerateResponse type');
-          }
-        } catch (error) {
-          aborted = true;
-          abortController.abort();
-          set(currentStreamAtom, null);
-          throw error;
+    async (
+        get,
+        set,
+        {
+            message,
+            response,
+            messages,
+            sources,
+            addMessageModifier,
+        }: {
+            message: string;
+            response?: Promise<string> | AsyncIterable<{ content: string; done?: boolean }>;
+            messages?: Promise<Message[]>;
+            sources?: Promise<Source[]>;
+            addMessageModifier?: AddUserMessageActionModifier;
         }
-      })();
+    ) => {
+        // Save previous state for rollback.
+        const previousMessages = get(completedMessagesAtom);
+        const previousSources = get(retrievedSourcesAtom);
 
-      /**
-       * Completion Tracking
-       * Manages loading state and error handling for the operation
-       */
-      processingPromise.catch(error => {
-        set(errorAtom, `Generate operation failed: ${error instanceof Error ? error.message : String(error)}`);
-        set(currentStreamAtom, null);
-      }).finally(() => {
-        set(loadingAtom, false);
-      });
+        // Read configuration and create a common AbortController.
+        const config = get(configAtom);
+        const abortController = new AbortController();
 
-      return response;
+        // Handle user message modification.
+        if (addMessageModifier?.setUserMessage) {
+            set(completedMessagesAtom, [
+                ...get(completedMessagesAtom),
+                {
+                    id: crypto.randomUUID() as UUID,
+                    role: 'user',
+                    content: addMessageModifier.setUserMessage,
+                },
+            ]);
+        } else {
+            set(completedMessagesAtom, [
+                ...get(completedMessagesAtom),
+                {
+                    id: crypto.randomUUID(),
+                    role: 'user',
+                    content: message,
+                },
+            ]);
+        }
+
+        // Process sources independently.
+        const processSources = async () => {
+            if (!sources) return; // Skip if no sources provided.
+            const sourcesData: Source[] = await addTimeout(
+                sources.then(srcs => srcs),
+                config.timeouts?.request,
+                'Sources request timeout exceeded',
+                abortController.signal
+            );
+
+            // Add a UUID to each source if it's missing.
+            const sourcesDataWithIds = sourcesData.map(source => ({
+                ...source,
+                id: source.id || (crypto.randomUUID() as UUID),
+            }));
+
+            // Update sources-related state.
+            set(retrievedSourcesAtom, sourcesDataWithIds);
+            set(activeSourcesIdsAtom, []);
+            set(selectedSourceIdAtom, null);
+            return sourcesDataWithIds;
+        };
+
+        // Process the user message or streaming response independently.
+        const processMessage = async () => {
+            if (response) {
+                let accumulatedContent = '';
+                if (Symbol.asyncIterator in response) {
+                    // Streaming response: process each chunk.
+                    const streamTimeout = new StreamTimeout(config.timeouts?.stream);
+                    for await (const chunk of response as AsyncIterable<{ content: string; done?: boolean }>) {
+                        if (abortController.signal.aborted) break;
+                        streamTimeout.check();
+                        accumulatedContent += chunk.content;
+                        // Provide immediate feedback as streaming chunks arrive.
+                        set(currentStreamAtom, {
+                            id: crypto.randomUUID(),
+                            role: 'assistant',
+                            content: accumulatedContent,
+                        });
+                    }
+                    // Finalize streaming.
+                    set(completedMessagesAtom, [
+                        ...get(completedMessagesAtom),
+                        {
+                            id: crypto.randomUUID() as UUID,
+                            role: 'assistant',
+                            content: accumulatedContent,
+                        },
+                    ]);
+                    set(currentStreamAtom, null);
+                    return accumulatedContent;
+                } else {
+                    // Non-streaming (single promise) response.
+                    const messageData = await addTimeout(
+                        response as Promise<string>,
+                        config.timeouts?.request,
+                        'Response timeout exceeded',
+                        abortController.signal
+                    );
+                    set(completedMessagesAtom, [
+                        ...get(completedMessagesAtom),
+                        {
+                            id: crypto.randomUUID() as UUID,
+                            role: 'assistant',
+                            content: messageData,
+                        } as Message,
+                    ]);
+                    return messageData;
+                }
+            }
+            if (messages) {
+                throw new Error('Not implemented yet. Should serve more agentic use cases lateron.');
+            }
+        };
+
+        // --- The key elegant change: since this function is async, it automatically returns a promise.
+        try {
+            const results = await Promise.allSettled([processSources(), processMessage()]);
+            const errors = results.filter(result => result.status === 'rejected');
+
+            if (errors.length > 0) {
+                // Abort remaining work and restore previous state.
+                abortController.abort();
+                set(completedMessagesAtom, previousMessages);
+                set(retrievedSourcesAtom, previousSources);
+                set(currentStreamAtom, null);
+
+                // Aggregate error messages.
+                const errorMessages = errors
+                    .map((e: PromiseRejectedResult) =>
+                        e.reason instanceof Error ? e.reason.message : String(e.reason)
+                    )
+                    .join('; ');
+                throw new Error(errorMessages);
+            }
+            return results;
+        } catch (error) {
+            // Notify about the error and propagate it.
+            set(setErrorAtom, `Failed to process user message: ${error instanceof Error ? error.message : error}`);
+            throw error;
+        }
     }
-  );
-};
+);
 
-// Retrieve sources
-export const retrievedSourcesAtom = atom<RetrievalResult[]>([]);
-export const currentSourceIndicesAtom = atom<number[]>([]);
 
-// Derive current sources from the selected indices
-export const currentSourcesAtom = atom((get) => {
-  const sources = get(retrievedSourcesAtom);
-  const indices = get(currentSourceIndicesAtom);
-
-  if (indices.length === 0) {
-    return [];
-  }
-
-  // Filter and sort indices to ensure they're valid and in order
-  const validIndices = indices
-    .filter(index => index >= 0 && index < sources.length)
-    .sort((a, b) => a - b);
-
-  // Return current sources in order
-  return validIndices.map(index => sources[index]);
+// set active message
+const setActiveMessageAtom = atom(null, (_get, set, { setActiveMessageModifier }: {
+    setActiveMessageModifier?: SetActiveMessageActionModifier
+}) => {
+    set(activeMessageIdAtom, setActiveMessageModifier?.messageId ?? null);
 });
 
-// Update setter atom to handle multiple indices
-export const setCurrentSourceIndicesAtom = atom(
-  null,
-  (_get, set, indices: number[]) => {
-    set(currentSourceIndicesAtom, indices);
-  }
-);
+// clear messages
+const clearMessagesAtom = atom(null, (_get, set, { clearMessagesModifier }: {
+    clearMessagesModifier?: ClearMessagesActionModifier
+}) => {
+    console.log('clearMessagesAtom', clearMessagesModifier);
+    set(completedMessagesAtom, []);
+});
 
-
-export const createRetrieveSourcesAtom = (retrieveFn: (query: string, metadata?: Record<string, any>) => Promise<RetrievalResult[]>) => {
-  return atom(
+// search sources
+const searchSourcesAtom = atom(
     null,
-    async (_get, set, query: string, metadata?: Record<string, any>) => {
-      set(loadingAtom, true);
-      try {
-        const sources = await retrieveFn(query, metadata);
-        set(retrievedSourcesAtom, sources);
-        set(currentSourceIndicesAtom, []); // Reset active source indices
-        set(activeSourceIndexAtom, null);  // Reset active source index
-        return sources;
-      } finally {
-        set(loadingAtom, false);
-      }
+    async (
+        get,
+        set,
+        { query, sources, searchSourcesModifier }: {
+            query: string,
+            sources?: Promise<Source[]>;
+            searchSourcesModifier?: SearchSourcesActionModifier
+        }
+    ) => {
+        // todo: make use of query and searchSourcesModifier
+        console.log('searchSourcesAtom', query, searchSourcesModifier);
+        // Save previous state for rollback.
+        const previousSources = get(retrievedSourcesAtom);
+
+        // Retrieve timeout configuration.
+        const config = get(configAtom);
+        const abortController = new AbortController();
+
+        // Set the loading state to block concurrent operations.
+        try {
+            // If no sources promise is provided, skip updating and return early.
+            if (!sources) {
+                console.warn("No sources provided to searchSourcesAtom. Operation skipped.");
+                return previousSources;
+            }
+
+            // Await the sources, applying the configured timeout.
+            const timedSources = await addTimeout(
+                sources,
+                config.timeouts?.request,
+                'Sources request timeout exceeded in searchSourcesAtom',
+                abortController.signal
+            );
+
+            // If successful, update the retrieved sources.
+            set(retrievedSourcesAtom, timedSources);
+
+            // Optionally reset related state associated with sources.
+            set(activeSourcesIdsAtom, []);
+            set(selectedSourceIdAtom, null);
+
+            return timedSources;
+        } catch (error) {
+            // Abort any further processing if an error occurs.
+            abortController.abort();
+
+            // Roll back to the previous state.
+            set(retrievedSourcesAtom, previousSources);
+
+            // Store the error message in the global error state.
+            set(
+                errorAtom,
+                `Failed to search sources: ${error instanceof Error ? error.message : String(error)}`
+            );
+
+            // Propagate the error for any higher-level handling.
+            throw error;
+        }
     }
-  );
-};
-
-export const retrieveSourcesAtom = atom(
-  null,
-  async (_get, set, query: string, metadata?: Record<string, any>) => {
-    // Validate processing state
-    if (_get(loadingAtom)) {
-      toast.error('RAG operation already in progress');
-      throw new Error('RAG operation already in progress');
-    }
-
-    // Get current state
-    const ragAtoms = _get(ragAtomsAtom);
-    const currentSources = _get(retrievedSourcesAtom);
-    const currentMode = _get(workflowModeAtom);
-
-    if (!ragAtoms) {
-      throw new Error('RAG atoms not initialized');
-    }
-
-    if (!ragAtoms.retrieveSourcesAtom) {
-      throw new Error('Retrieve sources not available');
-    }
-
-    // Prepare rollback state
-    const rollbackState = {
-      workflowMode: currentMode,
-      sources: currentSources
-    };
-
-    try {
-      // Trigger the retrieve operation
-      const retrieveSourcesAtom = ragAtoms.retrieveSourcesAtom as RetrieveSourcesAtom;
-      const sources = await set(retrieveSourcesAtom, query, metadata);
-      set(workflowModeAtom, 'follow-up');
-      return sources;
-    } catch (err: any) {
-      // Rollback state on error
-      set(workflowModeAtom, rollbackState.workflowMode);
-      set(retrievedSourcesAtom, rollbackState.sources);
-      set(errorAtom, `Retrieve sources operation failed: ${err.message}`);
-      throw err;
-    }
-  }
 );
 
-/**
- * Creates an atom that handles the RAG (Retrieval-Augmented Generation) process.
- * This atom manages both source retrieval and response generation independently
- * while maintaining proper error handling and state management.
- */
-export const createRetrieveAndGenerateAtom = (
-  retrieveAndGenerateFn: (query: GenerateInput, metadata?: Record<string, any>) => RetrieveAndGenerateResponse
-) => {
-  return atom(null, (_get, set, messages: GenerateInput, metadata?: Record<string, any>): RetrieveAndGenerateResponse => {
-    // Initialize state and control variables
-    set(loadingAtom, true);
-    set(errorAtom, null);
-    
-    let aborted = false;
-    const abortController = new AbortController();
-    const config = _get(ragConfigAtom);
-    let accumulatedContent = '';
+// clear sources
+const clearSourcesAtom = atom(null, (_get, set, { clearSourcesModifier }: {
+    clearSourcesModifier?: ClearSourcesActionModifier
+}) => {
+    console.log('clearSourcesAtom', clearSourcesModifier);
+    set(retrievedSourcesAtom, []);
+    set(activeSourcesIdsAtom, []);
+    set(selectedSourceIdAtom, null);
+});
 
-    // Get the response object from the provided function
-    // This contains potentially both sources and response promises
-    const response = retrieveAndGenerateFn(messages, metadata);
-
-    /**
-     * Source Processing
-     * Handles the retrieval and validation of sources independently.
-     * Updates UI state as soon as sources are available.
-     */
-    const wrappedSources = response.sources && addTimeout(
-      response.sources.then(sources => {
-        if (!aborted) {
-          // Update UI with validated sources
-          const validatedSources = validateRetrievalResults(sources);
-          set(retrievedSourcesAtom, validatedSources);
-          set(currentSourceIndicesAtom, []);
-          set(activeSourceIndexAtom, null);
-        }
-        return sources;
-      }),
-      config.timeouts?.request,
-      'Sources request timeout exceeded',
-      abortController.signal
-    );
-
-    /**
-     * Response Processing
-     * Handles the generation response independently from sources.
-     * Supports both streaming and non-streaming responses.
-     * Updates UI state as content becomes available.
-     */
-    const wrappedResponse = (async () => {
-      try {
-        if (Symbol.asyncIterator in response.response) {
-          // Handle streaming response
-          const processStream = async () => {
-            const streamTimeout = new StreamTimeout(config.timeouts?.stream);
-            
-            for await (const chunk of response.response as AsyncIterable<GenerateStreamChunk>) {
-              if (aborted || abortController.signal.aborted) {
-                break;
-              }
-
-              // Check for chunk timeout and update content
-              streamTimeout.check();
-              accumulatedContent += chunk.content;
-              
-              // Update UI with current stream
-              set(currentStreamAtom, { 
-                role: 'assistant', 
-                content: accumulatedContent 
-              });
-              
-              if (chunk.done) {
-                // On completion, update message history
-                set(completedMessagesAtom, prev => [...prev, { 
-                  role: 'assistant', 
-                  content: accumulatedContent 
-                }]);
-                set(currentStreamAtom, null);
-              }
-            }
-            return accumulatedContent;
-          };
-
-          // Add overall timeout for stream processing
-          await addTimeout(
-            processStream(),
-            config.timeouts?.request,
-            'Stream processing timeout exceeded',
-            abortController.signal
-          );
-          return accumulatedContent;
-        } else {
-          // Handle non-streaming response
-          const content = await addTimeout(
-            response.response as Promise<string>,
-            config.timeouts?.request,
-            'Response timeout exceeded',
-            abortController.signal
-          );
-          
-          // Update message history with complete response
-          set(completedMessagesAtom, prev => [...prev, { 
-            role: 'assistant', 
-            content 
-          }]);
-          set(currentStreamAtom, null);
-          return content;
-        }
-      } catch (error) {
-        aborted = true;
-        abortController.abort();
-        set(currentStreamAtom, null);
-        set(loadingAtom, false);
-        set(errorAtom, `Response processing failed: ${error instanceof Error ? error.message : String(error)}`);
-        throw error;
-      }
-    })();
-
-    /**
-     * Completion Tracking
-     * Manages loading state and error handling for both operations.
-     * Loading state is only cleared when both operations complete or on error
-     */
-    let responseComplete = false;
-    let sourcesComplete = !wrappedSources; // If no sources, mark as complete
-    let hasError = false;
-
-    // Handle sources completion
-    if (wrappedSources) {
-      wrappedSources.catch(error => {
-        hasError = true;
-        set(errorAtom, `Sources processing failed: ${error instanceof Error ? error.message : String(error)}`);
-        set(loadingAtom, false); // Reset loading state on error
-      }).finally(() => {
-        sourcesComplete = true;
-        if (responseComplete && !hasError) {
-          set(loadingAtom, false);
-        }
-      });
+// set active sources
+const setActiveSourcesAtom = atom(null, (_get, set, { sourceIds, setActiveSourcesModifier }: {
+    sourceIds?: UUID[],
+    setActiveSourcesModifier?: SetActiveSourcesActionModifier
+}) => {
+    const activeSourceIds = setActiveSourcesModifier?.activeSourceIds ?? sourceIds;
+    if (activeSourceIds) {
+        set(activeSourcesIdsAtom, activeSourceIds);
     }
+});
 
-    // Handle response completion
-    wrappedResponse.catch(error => {
-      hasError = true;
-      set(errorAtom, `Response processing failed: ${error instanceof Error ? error.message : String(error)}`);
-      set(loadingAtom, false); // Reset loading state on error
-    }).finally(() => {
-      responseComplete = true;
-      if (sourcesComplete && !hasError) {
-        set(loadingAtom, false);
-      }
-    });
+// set selected source
+const setSelectedSourceAtom = atom(null, (_get, set, { sourceId, setSelectedSourceModifier }: {
+    sourceId?: UUID,
+    setSelectedSourceModifier?: SetSelectedSourceActionModifier
+}) => {
+    const selectedSourceId = setSelectedSourceModifier?.selectedSourceId ?? sourceId;
+    if (selectedSourceId) {
+        set(selectedSourceIdAtom, selectedSourceId);
+    }
+});
 
-    // Return wrapped promises while maintaining the original interface
-    return {
-      sources: wrappedSources,
-      response: wrappedResponse
-    };
-  });
-};
+// set filter sources
+const setFilterSourcesAtom = atom(null, (_get, _set, { setFilterSourcesModifier }: {
+    setFilterSourcesModifier?: SetFilterSourcesActionModifier
+}) => {
+    console.log('setFilterSourcesAtom', setFilterSourcesModifier);
+    throw new Error('Not implemented yet');
+});
 
+// reset filter sources
+const resetFilterSourcesAtom = atom(null, (_get, _set, { resetFilterSourcesModifier }: {
+    resetFilterSourcesModifier?: ResetFilterSourcesActionModifier
+}) => {
+    console.log('resetFilterSourcesAtom', resetFilterSourcesModifier);
+    throw new Error('Not implemented yet');
+});
+
+
+// ----- MAIN: central dispatch atom / function -----
+export const dispatchAtom = atom(
+    null,
+    async (get, set, action: UserAction, recursiveCall: boolean = false) => {
+        // Exit if not a recursive call and we're already loading.
+        if (!recursiveCall && get(loadingAtom)) {
+            set(setErrorAtom, "RAG Operation already in progress");
+            return;
+        }
+
+        // Set the global loading state and clear error (only for top-level calls)
+        if (!recursiveCall) {
+            console.log("set loadingAtom to true");
+            set(loadingAtom, true);
+            set(errorAtom, null);
+        }
+
+        // Find the handler.
+        const handlers = get(registeredActionHandlersAtom);
+        // (For now we assume the RAGProvider is used; you might later use action.source.)
+        const handler = handlers.find(h => h.component === 'RAGProvider');
+        if (!handler) {
+            console.warn(`Handler for component ${action.source} not found`);
+            if (!recursiveCall) {
+                set(loadingAtom, false);
+            }
+            return;
+        }
+
+        // Call the handler and wrap its return value so that it always returns a promise.
+        // This handles both synchronous and asynchronous handlers.
+        const payload = await Promise.resolve(
+            handler.handler(
+                action,
+                get(completedMessagesAtom),
+                get(retrievedSourcesAtom),
+                get(activeSourcesAtom),
+                get(selectedSourceAtom)
+            )
+        );
+
+        // Validate payload keys.
+        const allowedKeys = allowedPayloadKeys[action.type] || [];
+        if (payload) {
+            const extraKeys = Object.keys(payload).filter(key => !allowedKeys.includes(key));
+            if (extraKeys.length > 0) {
+                console.warn(
+                    `Handler for action "${action.type}" returned unused properties: ${extraKeys.join(', ')}`
+                );
+            }
+        }
+
+        // Validate modifiers.
+        const { actionOptions } = payload ?? {};
+        if (actionOptions?.current) {
+            const allowedModifierKeys = Object.keys(allowedModifiers[action.type] || {});
+            const currentModifierKeys = Object.keys(actionOptions.current);
+            const extraModifierKeys = currentModifierKeys.filter(key => !allowedModifierKeys.includes(key));
+            if (extraModifierKeys.length > 0) {
+                console.warn(
+                    `Handler for action "${action.type}" used invalid modifiers: ${extraModifierKeys.join(', ')}`
+                );
+            }
+        }
+
+        // Destructure payload values.
+        const { response, messages, sources } = payload ?? {};
+
+        // Collect all write-atom calls in an array of promises.
+        // Each call is wrapped with Promise.resolve() so that synchronous and asynchronous atoms
+        // are handled uniformly.
+        const promises: Promise<any>[] = [];
+
+        switch (action.type) {
+            case 'ADD_USER_MESSAGE': {
+                const result = set(addUserMessageAtom, {
+                    message: action.message,
+                    response,
+                    messages,
+                    sources,
+                    addMessageModifier: actionOptions?.current as AddUserMessageActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'SET_ACTIVE_MESSAGE': {
+                const result = set(setActiveMessageAtom, {
+                    setActiveMessageModifier: actionOptions?.current as SetActiveMessageActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'CLEAR_MESSAGES': {
+                const result = set(clearMessagesAtom, {
+                    clearMessagesModifier: actionOptions?.current as ClearMessagesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'SEARCH_SOURCES': {
+                const result = set(searchSourcesAtom, {
+                    query: action.query,
+                    sources,
+                    searchSourcesModifier: actionOptions?.current as SearchSourcesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'CLEAR_SOURCES': {
+                const result = set(clearSourcesAtom, {
+                    clearSourcesModifier: actionOptions?.current as ClearSourcesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'SET_ACTIVE_SOURCES': {
+                const result = set(setActiveSourcesAtom, {
+                    sourceIds: action.sourceIds as UUID[],
+                    setActiveSourcesModifier: actionOptions?.current as SetActiveSourcesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'SET_SELECTED_SOURCE': {
+                const result = set(setSelectedSourceAtom, {
+                    sourceId: action.sourceId as UUID,
+                    setSelectedSourceModifier: actionOptions?.current as SetSelectedSourceActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'SET_FILTER_SOURCES': {
+                const result = set(setFilterSourcesAtom, {
+                    setFilterSourcesModifier: actionOptions?.current as SetFilterSourcesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            case 'RESET_FILTER_SOURCES': {
+                const result = set(resetFilterSourcesAtom, {
+                    resetFilterSourcesModifier: actionOptions?.current as ResetFilterSourcesActionModifier,
+                });
+                promises.push(Promise.resolve(result));
+                break;
+            }
+            default:
+                console.warn(`Unhandled action type: ${(action as any).type}`);
+        }
+
+        // Handle any follow-up actions recursively.
+        if (actionOptions?.followUp) {
+            const followUpResult = set(dispatchAtom, actionOptions.followUp, true);
+            promises.push(Promise.resolve(followUpResult));
+        }
+
+        // Wait for all write operations to finish.
+        try {
+            await Promise.all(promises);
+        } catch (error) {
+            console.error("Error in dispatch async writes:", error);
+        } finally {
+            if (!recursiveCall) {
+                console.log("set loadingAtom to false");
+                set(loadingAtom, false);
+            }
+        }
+    }
+);
+
+
+
+
+// ---- helpers for handling stream timeout and stream abortion ----
 const addTimeout = <T>(
-  promise: Promise<T>,
-  timeout?: number,
-  message?: string,
-  signal?: AbortSignal
+    promise: Promise<T>,
+    timeout?: number,
+    message?: string,
+    signal?: AbortSignal
 ): Promise<T> => {
-  if (!timeout) return promise;
+    if (!timeout) return promise;
 
-  return Promise.race([
-    promise,
-    new Promise<never>((_, reject) => {
-      const timeoutId = setTimeout(() => {
-        reject(new Error(message));
-      }, timeout);
+    return Promise.race([
+        promise,
+        new Promise<never>((_, reject) => {
+            const timeoutId = setTimeout(() => {
+                reject(new Error(message));
+            }, timeout);
 
-      // Clean up timeout if aborted
-      signal?.addEventListener('abort', () => {
-        clearTimeout(timeoutId);
-      });
-    })
-  ]);
+            // Clean up timeout if aborted
+            signal?.addEventListener('abort', () => {
+                clearTimeout(timeoutId);
+            });
+        })
+    ]);
 };
 
 // Helper classes/functions
 class StreamTimeout {
-  private lastCheck = Date.now();
+    private lastCheck = Date.now();
 
-  constructor(private timeout?: number) { }
-
-  check() {
-    if (!this.timeout) return;
-
-    const now = Date.now();
-    if (now - this.lastCheck > this.timeout) {
-      throw new Error('Stream timeout exceeded');
+    constructor(private timeout?: number) {
     }
-    this.lastCheck = now;
-  }
-}
 
-export const createGetDataSourceAtom = (getDataSourceFn: ((source: SourceReference) => GetDataSourceResponse) | null) => {
-  return atom<null, [RetrievalResult], Promise<SourceContent>>(
-    null,
-    (_get, set, retrievalResult: RetrievalResult): Promise<SourceContent> => {
-      set(loadingAtom, true);
-      set(errorAtom, null);
-      
-      try {
-        // Handle text content directly
-        if ('text' in retrievalResult) {
-          const response: SourceContent = {
-            content: retrievalResult.text,
-            metadata: retrievalResult.metadata,
-            type: 'markdown'
-          };
-          set(currentSourceContentAtom, response);
-          set(loadingAtom, false);
-          return Promise.resolve(response);
+    check() {
+        if (!this.timeout) return;
+
+        const now = Date.now();
+        if (now - this.lastCheck > this.timeout) {
+            throw new Error('Stream timeout exceeded');
         }
-
-        // If no getDataSourceFn is provided, throw an error for source references
-        if (!getDataSourceFn) {
-          const error = new Error('Source content retrieval is not configured. Only text sources are supported.');
-          set(errorAtom, error.message);
-          set(loadingAtom, false);
-          return Promise.reject(error);
-        }
-
-        // For SourceReference, use the getDataSourceFn
-        return getDataSourceFn(retrievalResult)
-          .then(response => {
-            // Set page number for PDFSourceContent if not provided
-            if (response.type && response.type === 'pdf' && !response.page) {
-              // Set page number from metadata if available
-              if (retrievalResult.metadata && retrievalResult.metadata?.page && typeof retrievalResult.metadata?.page === 'number') {
-                response.page = retrievalResult.metadata.page;
-              }
-            }
-            set(currentSourceContentAtom, response);
-            return response;
-          })
-          .finally(() => {
-            set(loadingAtom, false);
-          });
-      } catch (err) {
-        set(errorAtom, `Failed to fetch source content: ${err instanceof Error ? err.message : String(err)}`);
-        set(loadingAtom, false);
-        return Promise.reject(err);
-      }
+        this.lastCheck = now;
     }
-  );
-};
-
-type GenerateAtom = WritableAtom<null, [GenerateInput], GenerateResponse>;
-
-type RetrieveAndGenerateAtom = WritableAtom<
-  null,
-  [GenerateInput, Record<string, any>?],
-  RetrieveAndGenerateResponse
->;
-
-type RetrieveSourcesAtom = WritableAtom<
-  null,
-  [string, Record<string, any>?],
-  RetrieveResponse
->;
-
-type GetDataSourceAtom = WritableAtom<
-  null,
-  [RetrievalResult],
-  GetDataSourceResponse
->;
-
-// Container for all RAG atoms. This is necessary since the atoms are created dynamically but need to be referenced form other atoms
-interface RAGAtoms {
-  generateAtom: GenerateAtom | null;
-  retrieveAndGenerateAtom: RetrieveAndGenerateAtom | null;
-  retrieveSourcesAtom: RetrieveSourcesAtom | null;
-  getDataSourceAtom: GetDataSourceAtom | null;
 }
-
-export const ragAtomsAtom = atom<RAGAtoms | null>(null);
